101-print_number.c created

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -24,5 +24,5 @@ void print_number(int n)
 		print_number(i / 10);
 	}
 
-	_putchar((n % 10) + '0');
+	_putchar((unsigned int) n % 10 + '0');
 }
